requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0
scipy>=1.11.0
python-binance>=1.0.19
pyarrow>=16.0.0
